final/requirements.txt
# Frequency-Aware Vector Quantization for VLM Compression
# Python 3.10+ required

# Deep Learning
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.40.0

# Computer Vision & Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Numerical & Scientific Computing
numpy>=1.24.0
scipy>=1.11.0
matplotlib>=3.7.0

# Utilities
tqdm>=4.65.0
difflib  # Built-in Python module

# Note: CUDA 11.8+ required for GPU support
# Install PyTorch with CUDA from: https://pytorch.org/get-started/locally/
# For CUDA 11.8: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118

